pyproject.toml
[project]
name = "fala.ai"
version = "0.1.0"
description = "FALA.AI backend API"
authors = [{ name = "FALA.AI Team", email = "dev@fala.ai" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.110,<0.112",
    "uvicorn[standard]>=0.30,<0.31",
    "sqlalchemy[asyncio]>=2.0,<2.1",
    "asyncpg>=0.29,<0.30",
    "alembic>=1.13,<1.15",
    "python-jose[cryptography]>=3.3,<3.4",
    "passlib[bcrypt]>=1.7,<1.8",
    "pydantic-settings>=2.2,<2.3",
]

[project.optional-dependencies]
test = [
    "pytest>=8.2,<8.3",
    "pytest-asyncio>=0.23,<0.24",
    "httpx>=0.27,<0.28",
]

dev = [
    "ruff>=0.5.0,<0.6.0",
    "mypy>=1.10.0,<1.11.0",
]

[build-system]
requires = ["hatchling>=1.24,<2.0"]
build-backend = "hatchling.build"

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
    "app",
    "pyproject.toml",
    "README.md",
    "alembic",
    "docker",
    "scripts",
    "tests",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
select = ["E", "F", "I", "UP", "B", "SIM"]
line-length = 100

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
disable_error_code = ["has-type"]
strict_optional = true
